middleware::log_request: Split Sentry error reporting code into dedicated middleware
